chore: move parse_xxx to parser module
